+++ sdk/resourcemanager/eventhub/armeventhub/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armeventhub
 
 const (
 	moduleName    = "armeventhub"
-	moduleVersion = "v0.4.0"
+	moduleVersion = "v0.5.0"
 )
 
 type AccessRights string